Added basic handling of image functions
The read and write function have to be improved to actually load and
store shadow values for the images to shadow memory. At the moment only
the image object itself is checked. The correct handling would require
to copy all image related functions from WorkItemBuiltins.cpp into this
plugin which does not seem to be a good solution.