Simple plugin to detect arithmetic exceptions
It's capabilities are unfortunately limited by the fact that LLVM
doesn't make a difference between signed and unsigned operations in its
LLVM IR. To avoid excessive false positive warnings signed overflows are
only checked if the LLVM IR instructions contains the 'no sign wrap'
(nsw) flag. Vector operations seem no to include this flag.

Further the pointer to integer warning if the destination type is too
narrow to hold the pointer value is sometimes suppressed by the fact
that first a conversion from *iX to i64 is performed and only afterwards
a truncation from i64 to iY (destination type).