Implement program scope global variables
This is done generically for program scope variables in both the
constant and global address spaces. The lifetime of such variables is
with the program (whereas previously, the lifetime of constant globals
was with a kernel invocation).

Refactored GEP handling into a generic resolveGEP() in common.cpp, along
with a resolveConstantPointer() routine.