AIWC: don't store size of memory ops; fix bug in local memory entropy
Don't store the size of memory operations as it's not used in later analysis.
To determine the number of unique memory addresses, just use the keys of the
map that's used to store the count of accesses against each address.
In calculation of Local Memory Address Entropy, the address was being cast
from an unsigned to a signed int, which is incorrect for addresses where the
first bit is a one. Instead of doing a division (by float), just use a right
shift by the chosen number of bits.