Fix memory leaks when releasing commands
Instances of derived command structs with destructors were not
being freed since the Command base struct didn't have a virtual
destructor.

Found this when adding a destructor to the FillImageCommand
struct, and noticing that it was not being invoked.

I initially added this destructor due to an invalid read size
valgrind warning in the following instruction:

memcpy(color, col, 16);

Size of color object may not be 16 (it depends on pixel size),
thus the warning. As such, the color object is now dynamically
allocated depending on the pixel size, and destroyed in the
respective constructor.