Fix segfaults when modifying event type in image API functions
In Oclgrind some image OpenCL API functions use buffer functions
to complete their work. However, the image functions don't check
for errors thrown by the buffer functions and always try to modify
the corresponding event. Since an event is not created when the
buffer functions throw errors, the image functions try to modify
an invalid memory block, expecting it to contain an event,
resulting in a segmentation fault.

This fix checks if errors are thrown by buffer functions before
modifying the corresponding event in the calling image functions.