Fix removal of command from queue
Was not updating pointer to previous (older) command in next
(newer) command. This would be a problem if a command in the
middle of the queue was removed in a multiple queue or
out-of-order queue scenario.

The pointer to the previous (older) command has been removed,
so that command removal is done using iterators, which
simplifies the code compared with manually updating the pointer
to the previous command.

The only downside is that this is not very efficient, since
each command to be executed is always searched through the queue.